Create LESUtils module, reorganize src.
